@@ -35,40 +35,40 @@
   },
   "dependencies": {
     "@heroicons/react": "^2.0.18",
-    "@vitejs/plugin-react": "^1.2.0",
+    "@vitejs/plugin-react": "^4.1.0",
     "electron-is-dev": "^2.0.0",
     "electron-store": "^8.1.0",
     "eslint-plugin-neverthrow": "^1.1.4",
     "neverthrow": "^6.1.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hot-toast": "^2.4.1",
-    "react-router-dom": "^6.17.0",
-    "vite": "^2.8.6"
+    "react-router-dom": "^6.18.0",
+    "vite": "^4.5.0"
   },
   "devDependencies": {
-    "@types/babel__core": "^7.20.2",
-    "@types/node": "^20.8.3",
-    "@types/react": "^17.0.41",
-    "@types/react-dom": "^17.0.14",
-    "@typescript-eslint/eslint-plugin": "5.16.0",
-    "@typescript-eslint/parser": "5.16.0",
-    "autoprefixer": "^10.4.4",
-    "concurrently": "^7.0.0",
-    "electron": "^26.3.0",
+    "@types/babel__core": "^7.20.3",
+    "@types/node": "^20.8.10",
+    "@types/react": "^18.2.33",
+    "@types/react-dom": "^18.2.14",
+    "@typescript-eslint/eslint-plugin": "6.9.1",
+    "@typescript-eslint/parser": "6.9.1",
+    "autoprefixer": "^10.4.16",
+    "concurrently": "^8.2.2",
+    "electron": "^27.0.2",
     "electron-builder": "^24.6.4",
-    "eslint": "8.11.0",
+    "eslint": "8.52.0",
     "eslint-config-airbnb": "19.0.4",
-    "eslint-config-prettier": "8.5.0",
-    "eslint-plugin-import": "^2.25.4",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-prettier": "4.0.0",
-    "eslint-plugin-react": "^7.29.4",
-    "postcss": "^8.4.12",
-    "prettier": "2.6.0",
+    "eslint-config-prettier": "9.0.0",
+    "eslint-plugin-import": "^2.29.0",
+    "eslint-plugin-jsx-a11y": "^6.7.1",
+    "eslint-plugin-prettier": "5.0.1",
+    "eslint-plugin-react": "^7.33.2",
+    "postcss": "^8.4.31",
+    "prettier": "3.0.3",
     "simple-git-hooks": "^2.9.0",
-    "tailwindcss": "^3.0.23",
-    "typescript": "^4.6.2"
+    "tailwindcss": "^3.3.5",
+    "typescript": "^5.2.2"
   },
   "simple-git-hooks": {
     "pre-commit": "yarn type-check && yarn lint:fix"
